requirements.txt
flask
pynput
requests
pillow
mss
pyinstaller